requirements.txt
# Core dependencies for scripts
biopython==1.82
pandas==2.1.2
pyarrow==15.0.0
pyyaml==6.0.1
jsonschema==4.19.1
requests==2.31.0
matplotlib==3.8.0
seaborn==0.13.0
geopy==2.4.1
nextclade==2.14.0

# Advanced phylogenetic analysis dependencies
ete3>=3.1.2
dendropy>=4.5.2
numpy>=1.24.0
tqdm>=4.64.0

# Nextstrain-specific dependencies
nextstrain-cli>=5.0.0
augur>=21.1.0